添加代码，准备eval

+++ pytorch-pretrained-bert/script/run_absa.sh
@@ -36,7 +36,7 @@ do
             --bert_model $bert --do_eval --max_seq_length 100 \
             --output_dir $OUTPUT_DIR --data_dir $DATA_DIR --seed $run > $OUTPUT_DIR/test_log.txt 2>&1
     fi
-    if [ -e $OUTPUT_DIR/"predictions.json" ] && [ -e $OUTPUT_DIR/model.pt ] ; then
+    if [ -e $OUTPUT_DIR/"predictions.json" ] && [ -e $OUTPUT_DIR/model.pt ] && [[ run -ne runs ]]; then
         rm $OUTPUT_DIR/model.pt
     fi
 done
